Auto image radial blur for Spotify covers
The issue I was having before was that the app tried getting the cdCover and cdCoverFast too rapidly, while Python was not done blurring it. I fixed this by moving the loading of the blurred image later, when it actually needs it, allowing for a little more processing time, and it works. I also upped the speed a little on the closing/opening animation of the window